+++ src/types/panel.ts
@@ -82,9 +82,9 @@ export interface CalendarOptions {
   /**
    * Description
    *
-   * @type {string}
+   * @type {string[]}
    */
-  descriptionField?: string;
+  descriptionField?: string[];
 
   /**
    * End Time