+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Update variables with static datasource for time range source dashboard (#205)
+- Add multiple description fields (#206)
 
 ## 3.2.0 (2024-06-11)